# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "AutowatchRequest": {
    "additionalProperties": false,
    "properties": {
     "state": {
      "enum": [
       "running",
       "paused"
      ],
      "title": "State",
      "type": "string"
     },
     "until": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Until"
     }
    },
    "required": [
     "state"
    ],
    "title": "AutowatchRequest",
    "type": "object"
   },
   "AutowatchState": {
    "additionalProperties": false,
    "properties": {
     "state": {
      "enum": [
       "running",
       "paused"
      ],
      "title": "State",
      "type": "string"
     },
     "until": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Until"
     }
    },
    "required": [
     "state"
    ],
    "title": "AutowatchState",
    "type": "object"
   },
   "BookmarkIn": {
    "additionalProperties": false,
    "properties": {
     "actions": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "items": {},
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actions"
     },
     "check_method": {
      "$ref": "#/components/schemas/CheckMethod",
      "default": "auto"
     },
     "enabled": {
      "default": true,
      "title": "Enabled",
      "type": "boolean"
     },
     "fetch": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fetch"
     },
     "filter": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Filter"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "gate": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gate"
     },
     "highlight_mode": {
      "$ref": "#/components/schemas/HighlightMode",
      "default": "standard"
     },
     "info1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info1"
     },
     "info2": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info2"
     },
     "info3": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info3"
     },
     "macro_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Macro Id"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 300,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "plugin": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plugin"
     },
     "priority": {
      "default": 0,
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Priority",
      "type": "integer"
     },
     "schedule": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Schedule"
     },
     "source_type": {
      "$ref": "#/components/schemas/SourceType",
      "default": "auto"
     },
     "url": {
      "minLength": 1,
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "url"
    ],
    "title": "BookmarkIn",
    "type": "object"
   },
   "BookmarkOut": {
    "additionalProperties": false,
    "properties": {
     "actions": {
      "$ref": "#/components/schemas/ActionsConfig"
     },
     "baseline_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline Version Id"
     },
     "check_method": {
      "$ref": "#/components/schemas/CheckMethod"
     },
     "consecutive_errors": {
      "title": "Consecutive Errors",
      "type": "integer"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "current_interval_s": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Interval S"
     },
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     },
     "fetch": {
      "$ref": "#/components/schemas/FetchConfig"
     },
     "filter": {
      "$ref": "#/components/schemas/FilterConfig"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "gate": {
      "$ref": "#/components/schemas/GateConfig"
     },
     "gate_anchor_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gate Anchor Version Id"
     },
     "highlight_mode": {
      "$ref": "#/components/schemas/HighlightMode"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "info1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info1"
     },
     "info2": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info2"
     },
     "info3": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info3"
     },
     "keyword_hits": {
      "items": {
       "type": "string"
      },
      "title": "Keyword Hits",
      "type": "array"
     },
     "last_changed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Changed At"
     },
     "last_checked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Checked At"
     },
     "latest_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latest Version Id"
     },
     "macro_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Macro Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "next_due_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Due At"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "overrides": {
      "additionalProperties": true,
      "title": "Overrides",
      "type": "object"
     },
     "plugin": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plugin"
     },
     "priority": {
      "title": "Priority",
      "type": "integer"
     },
     "schedule": {
      "$ref": "#/components/schemas/ScheduleConfig"
     },
     "source_type": {
      "$ref": "#/components/schemas/SourceType"
     },
     "status": {
      "$ref": "#/components/schemas/BookmarkStatus"
     },
     "unread": {
      "title": "Unread",
      "type": "boolean"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "id",
     "folder_id",
     "name",
     "url",
     "source_type",
     "check_method",
     "enabled",
     "priority",
     "schedule",
     "fetch",
     "filter",
     "gate",
     "highlight_mode",
     "actions",
     "overrides",
     "macro_id",
     "plugin",
     "info1",
     "info2",
     "info3",
     "note",
     "status",
     "unread",
     "consecutive_errors",
     "current_interval_s",
     "next_due_at",
     "last_checked_at",
     "last_changed_at",
     "latest_version_id",
     "baseline_version_id",
     "gate_anchor_version_id",
     "created_at",
     "updated_at"
    ],
    "title": "BookmarkOut",
    "type": "object"
   },
...
   "BulkRequest": {
    "additionalProperties": false,
    "properties": {
     "action": {
      "enum": [
       "update",
       "move",
       "enable",
       "disable",
       "delete"
      ],
      "title": "Action",
      "type": "string"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "ids": {
      "items": {
       "type": "integer"
      },
      "minItems": 1,
      "title": "Ids",
      "type": "array"
     },
     "patch": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BookmarkPatch"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "ids",
     "action"
    ],
    "title": "BulkRequest",
    "type": "object"
   },
   "BulkResult": {
    "additionalProperties": false,
    "properties": {
     "affected": {
      "title": "Affected",
      "type": "integer"
     }
    },
    "required": [
     "affected"
    ],
    "title": "BulkResult",
    "type": "object"
   },
...
   "CheckQueued": {
    "additionalProperties": false,
    "properties": {
     "queued": {
      "title": "Queued",
      "type": "integer"
     }
    },
    "required": [
     "queued"
    ],
    "title": "CheckQueued",
    "type": "object"
   },
   "CheckRequest": {
    "additionalProperties": false,
    "properties": {
     "all": {
      "default": false,
      "title": "All",
      "type": "boolean"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "force": {
      "default": false,
      "title": "Force",
      "type": "boolean"
     },
     "ids": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ids"
     }
    },
    "title": "CheckRequest",
    "type": "object"
   },
...
   "FolderIn": {
    "additionalProperties": false,
    "properties": {
     "defaults": {
      "additionalProperties": true,
      "title": "Defaults",
      "type": "object"
     },
     "is_virtual": {
      "default": false,
      "title": "Is Virtual",
      "type": "boolean"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "query": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Query"
     },
     "sort_order": {
      "default": 0,
      "title": "Sort Order",
      "type": "integer"
     }
    },
    "required": [
     "name"
    ],
    "title": "FolderIn",
    "type": "object"
   },
   "FolderOut": {
    "additionalProperties": false,
    "properties": {
     "defaults": {
      "additionalProperties": true,
      "title": "Defaults",
      "type": "object"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "is_virtual": {
      "title": "Is Virtual",
      "type": "boolean"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     },
     "query": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Query"
     },
     "sort_order": {
      "title": "Sort Order",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "parent_id",
     "name",
     "sort_order",
     "is_virtual",
     "query",
     "defaults"
    ],
    "title": "FolderOut",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Page_BookmarkSummary_": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/BookmarkSummary"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[BookmarkSummary]",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "PageWatch engine",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/autowatch": {
   "get": {
    "operationId": "get_autowatch_autowatch_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutowatchState"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Autowatch",
    "tags": [
     "checks"
    ]
   },
   "post": {
    "operationId": "set_autowatch_autowatch_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AutowatchRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AutowatchState"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Set Autowatch",
    "tags": [
     "checks"
    ]
   }
  },
  "/bookmarks": {
   "get": {
    "operationId": "list_bookmarks_bookmarks_get",
    "parameters": [
     {
      "in": "query",
      "name": "folder",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Folder"
      }
     },
     {
      "in": "query",
      "name": "subfolders",
      "required": false,
      "schema": {
       "default": true,
       "title": "Subfolders",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "unread",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Unread"
      }
     },
     {
      "in": "query",
      "name": "enabled",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Enabled"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "id",
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "desc",
      "required": false,
      "schema": {
       "default": false,
       "title": "Desc",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_BookmarkSummary_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Bookmarks",
    "tags": [
     "bookmarks"
    ]
   },
   "post": {
    "operationId": "create_bookmark_bookmarks_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BookmarkIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Bookmark",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/bulk": {
   "post": {
    "operationId": "bulk_bookmarks_bulk_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BulkRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BulkResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bulk",
    "tags": [
     "bookmarks"
    ]
   }
  },
...
  "/bookmarks/{bookmark_id}/read": {
   "post": {
    "operationId": "mark_read_bookmarks__bookmark_id__read_post",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Mark Read",
    "tags": [
     "bookmarks"
    ]
   }
  },
...
  "/check": {
   "post": {
    "description": "Check a list of bookmark ids, a folder (with its subfolders), or everything.",
    "operationId": "check_check_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CheckRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CheckQueued"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Check",
    "tags": [
     "checks"
    ]
   }
  },
  "/folders": {
   "get": {
    "operationId": "list_folders_folders_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/FolderOut"
         },
         "title": "Response List Folders Folders Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Folders",
    "tags": [
     "folders"
    ]
   },
   "post": {
    "operationId": "create_folder_folders_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FolderIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FolderOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Folder",
    "tags": [
     "folders"
    ]
   }
  },
  "/folders/{folder_id}": {
   "delete": {
    "operationId": "delete_folder_folders__folder_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "folder_id",
      "required": true,
      "schema": {
       "title": "Folder Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Folder",
    "tags": [
     "folders"
    ]
   },
   "patch": {
    "operationId": "patch_folder_folders__folder_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "folder_id",
      "required": true,
      "schema": {
       "title": "Folder Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FolderPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FolderOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Folder",
    "tags": [
     "folders"
    ]
   }
  },
...
 }
}